Finish implementing
Codex prompt:

Implement the TODOs.

Also write a command called sctdemod which constructs a sctdemo.Server and calls Serve on it.  The server should be configured as follows:
GetCertificate should wrap GetCertificateFromFile from the src.agwa.name/go-listener/cert package, with the path specified using a command line flag
GetLog should scan the logs in a loglist.List (which was parsed from a JSON file specified as a command line flag) looking for a log whose hex-encoded LogID is prefixed by the argument passed to GetLog. If more than log has such a prefix, return an error.
THe cache functions should use a simple in-memory cache (with mutex protection).
The listener should be returned by Open from the src.agwa.name/go-listener package, with the listener spec passed in a command line flag

Codex worked for 7m 41s while I ate breakfast. I don't love the code but it does appear to be correct and good enough.